+++ src/basic/main.basic.js
@@ -62,8 +62,10 @@ const productList = [
 
 let cartDisp;
 
+import { renderCartList } from '../components/CartList.js';
 import { renderHeader } from '../components/Header.js';
 import { renderManualOverlay } from '../components/ManualOverlay.js';
+import { renderOrderSummary } from '../components/OrderSummary.js';
 
 function main() {
   // Header 분리된 컴포넌트 사용
@@ -71,7 +73,8 @@ function main() {
   const gridContainer = document.createElement('div');
   const leftColumn = document.createElement('div');
   const selectorContainer = document.createElement('div');
-  const rightColumn = document.createElement('div');
+  // OrderSummary 분리된 컴포넌트 사용
+  const rightColumn = renderOrderSummary();
   // ManualOverlay 분리된 컴포넌트 사용
   const { manualToggle, manualOverlay } = renderManualOverlay();
   totalAmt = 0;
@@ -98,38 +101,7 @@ function main() {
   selectorContainer.appendChild(stockInfo);
   leftColumn.appendChild(selectorContainer);
   cartDisp = document.createElement('div');
-  leftColumn.appendChild(cartDisp);
-  cartDisp.id = 'cart-items';
-  rightColumn.className = 'bg-black text-white p-8 flex flex-col';
-  rightColumn.innerHTML = `
-    <h2 class="text-xs font-medium mb-5 tracking-extra-wide uppercase">Order Summary</h2>
-    <div class="flex-1 flex flex-col">
-      <div id="summary-details" class="space-y-3"></div>
-      <div class="mt-auto">
-        <div id="discount-info" class="mb-4"></div>
-        <div id="cart-total" class="pt-5 border-t border-white/10">
-          <div class="flex justify-between items-baseline">
-            <span class="text-sm uppercase tracking-wider">Total</span>
-            <div class="text-2xl tracking-tight">₩0</div>
-          </div>
-          <div id="loyalty-points" class="text-xs text-blue-400 mt-2 text-right">적립 포인트: 0p</div>
-        </div>
-        <div id="tuesday-special" class="mt-4 p-3 bg-white/10 rounded-lg hidden">
-          <div class="flex items-center gap-2">
-            <span class="text-2xs">🎉</span>
-            <span class="text-xs uppercase tracking-wide">Tuesday Special 10% Applied</span>
-          </div>
-        </div>
-      </div>
-    </div>
-    <button class="w-full py-4 bg-white text-black text-sm font-normal uppercase tracking-super-wide cursor-pointer mt-6 transition-all hover:-translate-y-0.5 hover:shadow-lg hover:shadow-black/30">
-      Proceed to Checkout
-    </button>
-    <p class="mt-4 text-2xs text-white/60 text-center leading-relaxed">
-      Free shipping on all orders.<br>
-      <span id="points-notice">Earn loyalty points with purchase.</span>
-    </p>
-  `;
+  leftColumn.appendChild(renderCartList({ cartDisp }));
   sum = rightColumn.querySelector('#cart-total');
   gridContainer.appendChild(leftColumn);
   gridContainer.appendChild(rightColumn);
