+++ src/components/CartList.js
@@ -1,0 +1,7 @@
+// CartList.js
+// 장바구니 목록 UI 컴포넌트 분리
+
+export function renderCartList({ cartDisp }) {
+  cartDisp.id = 'cart-items';
+  return cartDisp;
+}
